fix reselecting same customer on queue edit
normally, when queue is initially completed with 'account balance' as
the payment option while customer's balance is lower than the grand
total price, the payment method should remains consistent to keep
selects 'account balance' option.

however, if the same customer was removed and reselected again or status
was changed, the payment method would change to 'cash' and
'account balance' option will be disabled. this commit ensures that
the 'account balance' option remains selectable as intended.