implement a proper way to manage fragment state and arguments
this commit fixes the previous behavior of argument retrieval.
previously, there was no way to check if the fragment was actually newly
created or if it was previously in the backstack — where the last
fragment was popped — to determine whether the fragment state should be
restored or not. by introducing Hilt, `SavedStateHandle` is
automatically injected during view model creation. from this we can
safely say that the fragment has just been newly created.