ref: simplify the way to auto update model in `LiveData`
while the static methods in `ModelUpdater` are slightly more verbose,
this approach clarifies behavior and centralizes data modification
logic, making it easier to understand and maintain.

previously, updating the data before serving it to `LiveData` was
harder due to `ModelChangedListener` returning void. this refactor
simplifies the process and improves the flexibility of data handling
before updating `LiveData`.