fix: bar chart aren't grouped by year when the `all time` range is set
this issue arises because the previous method heavily relies on
counting dates between the start and end dates (`totalDate`). however,
when the total number of months is less than 12, and the start and end
dates are in different years, the logic fails to group them by year
and instead groups them by month.