fix: improve debt filter to be more intuitive
customer debt is always in negative value. previously, when the debt
was $0, if the user input a minimum amount like -$1, they would expect
the customer to be filtered out, but it wasn't because -$1 is less than
$0. this change also ensures that users no longer need to consider
whether to use a negative prefix, as both options ($1 or -$1) will now
work the same.