fix: automatically adjust margin for tick labels on axis
by default, d3.js doesn't provide any margin for the text on the axis,
causing the text to be clipped. previously, the margin was set by a
hardcoded value, which isn't ideal when the labels have varying widths.
this could result in excessive empty space when the text is small.

this commit addresses these issues by implementing a
`_measureSpaceForTicksLabel()` function that temporarily renders the
text to measure it for the actual drawing later. however,
the measurement can only occur once the SVG is drawn. therefore, any
calculations regarding margin will take place in the render function
itself. in short, the `scale.range()` and `axis.tickSize()` have been
moved out of the `createXScale()` function.